Add files via upload
This code takes in Trackman files and calculates a pitcher's average metrics for their fastball, along with various results metrics such as Strike%, Whiff%, and Barrel% allowed. It also splits pitchers into buckets based on handedness, fastball type, and Extension to create a percentile for the aforementioned metrics and results.